// eval-fieldops/out-fieldops/api/db/repositories/organization-repository.ts
type Db = NodePgDatabase<typeof schema>;

export class OrganizationRepository implements OrganizationRepositoryPort {
  private readonly db: Db;
  private readonly events: DomainEventDispatcher;
  constructor(
    db: Db,
    events: DomainEventDispatcher,
  
// ...
}