// eval-fieldops/out-fieldops/api/db/repositories/asset-repository.ts
type Db = NodePgDatabase<typeof schema>;

export class AssetRepository implements AssetRepositoryPort {
  private readonly db: Db;
  private readonly events: DomainEventDispatcher;
  constructor(
    db: Db,
    events: DomainEventDispatcher,
  
// ...
}